# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClearPay Sales System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reps:
    get:
      operationId: listReps
      tags: [reps]
      summary: List all sales reps
      responses:
        "200":
          description: List of reps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rep"
    post:
      operationId: createRep
      tags: [reps]
      summary: Create a new rep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepBody"
      responses:
        "201":
          description: Created rep
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rep"
  /reps/{id}:
    get:
      operationId: getRep
      tags: [reps]
      summary: Get a rep by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rep
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rep"
        "404":
          description: Not found
    patch:
      operationId: updateRep
      tags: [reps]
      summary: Update a rep
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRepBody"
      responses:
        "200":
          description: Updated rep
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rep"
    delete:
      operationId: deleteRep
      tags: [reps]
      summary: Delete a rep
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /reps/{id}/stats:
    get:
      operationId: getRepStats
      tags: [reps]
      summary: Get performance stats for a rep
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rep stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepStats"
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - in: query
          name: assignedTo
          schema:
            type: ["integer", "null"]
        - in: query
          name: completed
          schema:
            type: boolean
        - in: query
          name: priority
          schema:
            type: string
        - in: query
          name: dueToday
          schema:
            type: boolean
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: ["integer", "null"]
        leadName:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        priority:
          type: string
        completed:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - priority
        - completed
        - createdAt
    CreateTaskBody:
      type: object
      properties:
        leadId:
          type: ["integer", "null"]
        assignedTo:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        priority:
          type: string
      required:
        - title
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        priority:
          type: string
        completed:
          type: boolean
        assignedTo:
          type: ["integer", "null"]
    Rep:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
    CreateRepBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
      required:
        - name
        - email
    UpdateRepBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
    RepStats:
      type: object
      properties:
        repId:
          type: integer
        totalLeads:
          type: integer
        closedWon:
          type: integer
        closedLost:
          type: integer
        inPipeline:
          type: integer
        conversionRate:
          type: number
        totalVolume:
          type: number
      required:
        - repId
        - totalLeads
        - closedWon
        - closedLost
        - inPipeline
        - conversionRate
        - totalVolume
...