# lib/api-zod/src/generated/types/createRepBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClearPay Sales System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRepBody {
  name: string;
  email: string;
  role?: string;
  /** @nullable */
  avatarUrl?: string | null;
}